Move Post API actions to their own file
